feature: Add !for! and !settings! commands

+++ regenerate-compilation-failures.sh
@@ -5,4 +5,6 @@ set -e
 TRYBUILD=overwrite cargo test
 
 # Remove the .wip folder created sometimes by try-build
-rm -r ./wip
+if [ -d "./wip" ]; then
+  rm -r ./wip
+fi
